+++ back-end/requests/webhook_trigger.rest
@@ -1,5 +1,5 @@
 # this can be any method
-POST http://localhost:3000/api/request/eca209ac-e719-4d08-88af-da33e46f0e7c
+POST http://localhost:5173/api/request/424e4b2e-fcb5-471f-a530-d52097849f8e
 
 Content-Type: application/json
