+++ front-end/package.json
@@ -11,6 +11,8 @@
   },
   "dependencies": {
     "axios": "^1.7.2",
+    "moment": "^2.30.1",
+    "moment-timezone": "^0.5.45",
     "react": "^18.2.0",
     "react-dom": "^18.2.0"
   },
