+++ front-end/src/Components/RequestList.jsx
@@ -1,4 +1,6 @@
 /* eslint-disable react/prop-types */
+// import moment from "moment";
+// import "moment-timezone";
 
 // Populates LHS list
 // Passes the MongoDB ID when clicked
@@ -9,12 +11,7 @@ const RequestList = ({ requests, handleRequestClick }) => {
 
   const requestInfo = ({ request_id, created_at, method, path }) => {
     const displayTime = () => {
-      let time = created_at.toString().substring(11, 19);
-      if (Number(time.substring(0, 2)) > 12) {
-        return Number(time.substring(0, 2)) - 12 + time.substring(2) + " PM";
-      } else {
-        return time + " AM";
-      }
+      return new Date(created_at.replace(" ", "T")).toString().substring(0, 24);
     };
     return (
       <div
