+++ front-end/src/services/requests.js
@@ -19,17 +19,17 @@ const getAllByToken = async (webhookToken) => {
     // console.log(response.data)
 
     if (response.data) {
-      let index = 0;
-      const currentDate = new Date();
-      while (index < response.data.length) {
-        if (currentDate - new Date(response.data[index].created_at) > 30000) {
-          await deleteById(response.data[index].request_id);
-          response.data.splice(index, 1);
-          continue;
-        }
+      // let index = 0;
+      // const currentDate = new Date();
+      // while (index < response.data.length) {
+      //   if (currentDate - new Date(response.data[index].created_at) > 30000) {
+      //     await deleteById(response.data[index].request_id);
+      //     response.data.splice(index, 1);
+      //     continue;
+      //   }
 
-        index += 1;
-      }
+      //   index += 1;
+      // }
       return response.data;
     } else {
       throw "Webhook Token Not Found";
